+++ examples/apollo-federation/helpers/buildFederatedSchema.ts
@@ -1,25 +1,24 @@
-import { specifiedDirectives } from "graphql";
-import federationDirectives from "@apollo/federation/dist/directives";
-import gql from "graphql-tag";
-import {
-  printSchema,
-  buildFederatedSchema as buildApolloFederationSchema,
-} from "@apollo/federation";
-import { addResolversToSchema, GraphQLResolverMap } from "apollo-graphql";
 import { buildSchema, BuildSchemaOptions, createResolversMap } from "../../../src";
+import { specifiedDirectives } from "graphql";
+import gql from "graphql-tag";
+import { buildSubgraphSchema } from "@apollo/subgraph";
+import { addResolversToSchema } from "@graphql-tools/schema";
+import { IResolvers, printSchemaWithDirectives } from "@graphql-tools/utils";
 
 export async function buildFederatedSchema(
   options: Omit<BuildSchemaOptions, "skipCheck">,
-  referenceResolvers?: GraphQLResolverMap<any>,
+  referenceResolvers?: IResolvers,
 ) {
   const schema = await buildSchema({
     ...options,
-    directives: [...specifiedDirectives, ...federationDirectives, ...(options.directives || [])],
+    directives: [...specifiedDirectives, ...(options.directives || [])],
     skipCheck: true,
   });
 
-  const federatedSchema = buildApolloFederationSchema({
-    typeDefs: gql(printSchema(schema)),
+  console.log(printSchemaWithDirectives(schema));
+
+  const federatedSchema = buildSubgraphSchema({
+    typeDefs: gql(printSchemaWithDirectives(schema)),
     resolvers: createResolversMap(schema) as any,
   });