+++ examples/custom-validation/recipes-arguments.ts
@@ -5,11 +5,11 @@ import { ArgsType, Field, Int } from "../../src";
 export class RecipesArguments {
   @Field(type => Int)
   // use decorators for Joi
-  @(Joiful.number().min(0))
+  @Joiful.number().min(0)
   skip: number = 0;
 
   @Field(type => Int)
   // use decorators for Joi
-  @(Joiful.number().min(1).max(50))
+  @Joiful.number().min(1).max(50)
   take: number = 10;
 }