+++ examples/extensions/helpers/config.extractors.ts
@@ -1,9 +1,8 @@
 import { GraphQLResolveInfo, GraphQLFieldConfig, GraphQLObjectTypeConfig } from "graphql";
 
 export const extractFieldConfig = (info: GraphQLResolveInfo): GraphQLFieldConfig<any, any> => {
-  const { type, extensions, description, deprecationReason } = info.parentType.getFields()[
-    info.fieldName
-  ];
+  const { type, extensions, description, deprecationReason } =
+    info.parentType.getFields()[info.fieldName];
 
   return {
     type,