+++ examples/graphql-modules/index.ts
@@ -18,7 +18,6 @@ async function bootstrap() {
 
   const server = new ApolloServer({
     schema,
-    playground: true,
   });
 
   const { url } = await server.listen(4000);