+++ tests/functional/authorization.ts
@@ -203,7 +203,7 @@ describe("Authorization", () => {
         normalQuery
       }`;
 
-      const result = await graphql(schema, query);
+      const result: any = await graphql({ schema, source: query });
 
       expect(result.data!.normalQuery).toEqual(true);
     });
@@ -215,7 +215,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result = await graphql(schema, query);
+      const result: any = await graphql({ schema, source: query });
 
       expect(result.data!.normalObjectQuery.normalField).toEqual("normalField");
     });
@@ -227,7 +227,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result = await graphql(schema, query);
+      const result: any = await graphql({ schema, source: query });
 
       expect(result.data!.normalObjectQuery.normalResolvedField).toEqual("normalResolvedField");
     });
@@ -241,7 +241,7 @@ describe("Authorization", () => {
         authedQuery
       }`;
 
-      const result = await graphql(localSchema, query);
+      const result: any = await graphql({ schema: localSchema, source: query });
 
       expect(result.data).toBeNull();
       expect(result.errors).toBeDefined();
@@ -257,7 +257,7 @@ describe("Authorization", () => {
         nullableAuthedQuery
       }`;
 
-      const result = await graphql(localSchema, query);
+      const result: any = await graphql({ schema: localSchema, source: query });
 
       expect(result.data!.nullableAuthedQuery).toBeNull();
       expect(result.errors).toBeUndefined();
@@ -272,7 +272,7 @@ describe("Authorization", () => {
         authedQuery
       }`;
 
-      const result = await graphql(localSchema, query);
+      const result: any = await graphql({ schema: localSchema, source: query });
 
       expect(result.data).toBeNull();
       expect(result.errors).toHaveLength(1);
@@ -290,7 +290,7 @@ describe("Authorization", () => {
         adminQuery
       }`;
 
-      const result = await graphql(localSchema, query);
+      const result: any = await graphql({ schema: localSchema, source: query });
 
       expect(result.data).toBeNull();
       expect(result.errors).toHaveLength(1);
@@ -308,7 +308,7 @@ describe("Authorization", () => {
         authedQuery
       }`;
 
-      const result = await graphql(localSchema, query, null, {});
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data!.authedQuery).toEqual(false);
     });
@@ -324,7 +324,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result = await graphql(localSchema, query);
+      const result: any = await graphql({ schema: localSchema, source: query });
 
       expect(result.data).toBeNull();
     });
@@ -340,7 +340,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result = await graphql(localSchema, query);
+      const result: any = await graphql({ schema: localSchema, source: query });
 
       expect(result.data!.normalObjectQuery.nullableAuthedField).toBeNull();
     });
@@ -356,7 +356,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result = await graphql(localSchema, query);
+      const result: any = await graphql({ schema: localSchema, source: query });
 
       expect(result.data).toBeNull();
       expect(result.errors).toHaveLength(1);
@@ -376,7 +376,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result = await graphql(localSchema, query);
+      const result: any = await graphql({ schema: localSchema, source: query });
 
       expect(result.data).toBeNull();
       expect(result.errors).toHaveLength(1);
@@ -396,7 +396,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result = await graphql(localSchema, query, null, {});
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data!.normalObjectQuery.authedField).toEqual("authedField");
     });
@@ -412,7 +412,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result = await graphql(localSchema, query);
+      const result: any = await graphql({ schema: localSchema, source: query });
 
       expect(result.data).toBeNull();
     });
@@ -428,7 +428,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result = await graphql(localSchema, query);
+      const result: any = await graphql({ schema: localSchema, source: query });
 
       expect(result.data).toBeNull();
     });
@@ -444,7 +444,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result = await graphql(localSchema, query, null, {});
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data!.normalObjectQuery.inlineAuthedResolvedField).toEqual(
         "inlineAuthedResolvedField",
@@ -464,7 +464,7 @@ describe("Authorization", () => {
         adminOrRegularQuery
       }`;
 
-      const result = await graphql(localSchema, query, null, {});
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data!.adminOrRegularQuery).toEqual(false);
       expect(authCheckerRoles).toEqual(["ADMIN", "REGULAR"]);
@@ -483,12 +483,12 @@ describe("Authorization", () => {
         adminOrRegularQuery
       }`;
 
-      const result = await graphql(
-        localSchema,
-        query,
-        { field: "rootField" },
-        { field: "contextField" },
-      );
+      const result: any = await graphql({
+        schema: localSchema,
+        source: query,
+        rootValue: { field: "rootField" },
+        contextValue: { field: "contextField" },
+      });
 
       expect(result.data!.adminOrRegularQuery).toEqual(false);
       expect(authCheckerResolverData.root.field).toEqual("rootField");
@@ -520,12 +520,12 @@ describe("Authorization", () => {
         }
       `;
 
-      const result = await graphql(
-        localSchema,
-        query,
-        { field: "rootField" },
-        { field: "contextField" },
-      );
+      const result: any = await graphql({
+        schema: localSchema,
+        source: query,
+        rootValue: { field: "rootField" },
+        contextValue: { field: "contextField" },
+      });
 
       expect(result.data).toBeNull();
       expect(result.errors).toMatchInlineSnapshot(`