+++ tests/functional/default-nullable.ts
@@ -85,8 +85,10 @@ describe("buildSchema -> nullableByDefault", () => {
       const normalArrayField = sampleObjectType.fields.find(it => it.name === "normalArrayField");
       const normalArrayFieldType = normalArrayField!.type as IntrospectionNonNullTypeRef;
       const normalArrayFieldListType = normalArrayFieldType.ofType as IntrospectionListTypeRef;
-      const normalArrayFieldListElementType = normalArrayFieldListType.ofType as IntrospectionNonNullTypeRef;
-      const normalArrayFieldListElementInnerType = normalArrayFieldListElementType.ofType as IntrospectionNamedTypeRef;
+      const normalArrayFieldListElementType =
+        normalArrayFieldListType.ofType as IntrospectionNonNullTypeRef;
+      const normalArrayFieldListElementInnerType =
+        normalArrayFieldListElementType.ofType as IntrospectionNamedTypeRef;
       expect(normalArrayFieldType.kind).toBe(TypeKind.NON_NULL);
       expect(normalArrayFieldListType.kind).toBe(TypeKind.LIST);
       expect(normalArrayFieldListElementType.kind).toBe(TypeKind.NON_NULL);
@@ -104,8 +106,10 @@ describe("buildSchema -> nullableByDefault", () => {
       const normalArrayQuery = queryType.fields.find(it => it.name === "normalArrayQuery");
       const normalArrayQueryType = normalArrayQuery!.type as IntrospectionNonNullTypeRef;
       const normalArrayQueryListType = normalArrayQueryType.ofType as IntrospectionListTypeRef;
-      const normalArrayQueryListElementType = normalArrayQueryListType.ofType as IntrospectionNonNullTypeRef;
-      const normalArrayQueryListElementInnerType = normalArrayQueryListElementType.ofType as IntrospectionNamedTypeRef;
+      const normalArrayQueryListElementType =
+        normalArrayQueryListType.ofType as IntrospectionNonNullTypeRef;
+      const normalArrayQueryListElementInnerType =
+        normalArrayQueryListElementType.ofType as IntrospectionNamedTypeRef;
       expect(normalArrayQueryType.kind).toBe(TypeKind.NON_NULL);
       expect(normalArrayQueryListType.kind).toBe(TypeKind.LIST);
       expect(normalArrayQueryListElementType.kind).toBe(TypeKind.NON_NULL);
@@ -137,7 +141,8 @@ describe("buildSchema -> nullableByDefault", () => {
 
       const normalArrayField = sampleObjectType.fields.find(it => it.name === "normalArrayField");
       const normalArrayFieldType = normalArrayField!.type as IntrospectionListTypeRef;
-      const normalArrayFieldListElementInnerType = normalArrayFieldType.ofType as IntrospectionNamedTypeRef;
+      const normalArrayFieldListElementInnerType =
+        normalArrayFieldType.ofType as IntrospectionNamedTypeRef;
       expect(normalArrayFieldType.kind).toBe(TypeKind.LIST);
       expect(normalArrayFieldListElementInnerType.kind).toBe(TypeKind.SCALAR);
       expect(normalArrayFieldListElementInnerType.name).toBe("String");
@@ -150,7 +155,8 @@ describe("buildSchema -> nullableByDefault", () => {
 
       const normalArrayQuery = queryType.fields.find(it => it.name === "normalArrayQuery");
       const normalArrayQueryType = normalArrayQuery!.type as IntrospectionListTypeRef;
-      const normalArrayQueryListElementInnerType = normalArrayQueryType.ofType as IntrospectionNamedTypeRef;
+      const normalArrayQueryListElementInnerType =
+        normalArrayQueryType.ofType as IntrospectionNamedTypeRef;
       expect(normalArrayQueryType.kind).toBe(TypeKind.LIST);
       expect(normalArrayQueryListElementInnerType.kind).toBe(TypeKind.SCALAR);
       expect(normalArrayQueryListElementInnerType.name).toBe("String");